put page objct at frontend lib [skip ci]
