+++ frontend/src/component/Workspace/ShareFolder.jsx
@@ -5,8 +5,8 @@ import { translate } from 'react-i18next'
 import classnames from 'classnames'
 import ContentItem from './ContentItem.jsx'
 import Folder from './Folder.jsx'
-import { PAGE, SHARE_FOLDER_ID } from '../../util/helper.js'
-import { ROLE, DropdownMenu } from 'tracim_frontend_lib'
+import { SHARE_FOLDER_ID } from '../../util/helper.js'
+import { ROLE, DropdownMenu, PAGE } from 'tracim_frontend_lib'
 
 require('./Folder.styl')
