+++ frontend/src/container/Login.jsx
@@ -15,6 +15,7 @@ import {
   CUSTOM_EVENT,
   NUMBER_RESULTS_BY_PAGE,
   checkEmailValidity,
+  PAGE,
   serialize
 } from 'tracim_frontend_lib'
 import {
@@ -47,10 +48,7 @@ import {
   putUserLang,
   getAccessibleWorkspaces
 } from '../action-creator.async.js'
-import {
-  PAGE,
-  COOKIE_FRONTEND
-} from '../util/helper.js'
+import { COOKIE_FRONTEND } from '../util/helper.js'
 import { serializeUserProps } from '../reducer/user.js'
 
 const qs = require('query-string')
