+++ frontend_app_folder_advanced/src/container/FolderAdvanced.jsx
@@ -16,6 +16,7 @@ import {
   ROLE,
   buildHeadTitle,
   BREADCRUMBS_TYPE,
+  PAGE,
   RefreshWarningMessage,
   TLM_ENTITY_TYPE as TLM_ET,
   TLM_CORE_EVENT_TYPE as TLM_CET,
@@ -187,7 +188,7 @@ export class FolderAdvanced extends React.Component {
       type: CUSTOM_EVENT.APPEND_BREADCRUMBS,
       data: {
         breadcrumbs: [{
-          url: `/ui/workspaces/${content.workspace_id}/contents/${state.config.slug}/${content.content_id}`,
+          url: PAGE.WORKSPACE.CONTENT(state.content.workspace_id, state.config.slug, state.content.content_id),
           label: content.label,
           link: null,
           type: BREADCRUMBS_TYPE.APP_FEATURE
