+++ frontend_app_thread/src/container/Thread.jsx
@@ -27,6 +27,7 @@ import {
   TracimComponent,
   getOrCreateSessionClientToken,
   getContentComment,
+  PAGE,
   permissiveNumberEqual
 } from 'tracim_frontend_lib'
 import {
@@ -264,7 +265,7 @@ export class Thread extends React.Component {
       type: CUSTOM_EVENT.APPEND_BREADCRUMBS,
       data: {
         breadcrumbs: [{
-          url: `/ui/workspaces/${content.workspace_id}/contents/${state.config.slug}/${content.content_id}`,
+          url: PAGE.WORKSPACE.CONTENT(state.content.workspace_id, state.config.slug, state.content.content_id),
           label: content.label,
           link: null,
           type: BREADCRUMBS_TYPE.APP_FEATURE
